强行修复dev上个版本造成的运行错误

+++ app/Http/Controllers/BookRequestController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\UserInfo;
 use Illuminate\Http\Request;
 
 use App\Models\User;
@@ -26,9 +27,10 @@ class BookRequestController extends Controller
 
     public function index(Request $request){
     	$user = Auth::user();
+        $userinfo = UserInfoController::get_user_info($user);
     	if(!empty($user->json_content))
     	    $user->json_content = json_decode($user->json_content);
-		return view('book_request.index',['user'=>$user]);
+		return view('book_request.index', ['user'=>$user, 'userinfo'=>$userinfo]);
     }
 
     public function storeMultiple(Request $request){
@@ -60,7 +62,7 @@ class BookRequestController extends Controller
 
 	    if( $user_json['teacher']['book_limit'] - sizeof($book_ids) < 0){
 		    $request->session()->flash('notice_status','danger');
-		    $request->session()->flash('notice_message','您申请的书籍多余您的限额，请重新申请!');
+		    $request->session()->flash('notice_message','您申请的书籍多于您的限额，请重新申请!');
 		    return redirect()->route('bookreq.index');
 	    }
 
@@ -80,11 +82,7 @@ class BookRequestController extends Controller
 	    	$book_req->save();
 	    }
 
-	    $user_json['address'] = [
-		    'receiver' => $receiver,
-		    'location'  => $address,
-		    'phone'    => $phone
-	    ];
+	    UserInfo::where('user_id',$user->id)->update(['address'=>$address]);
 
 	    $user_json['teacher']['book_limit'] -= sizeof($book_ids);
 	    $user->json_content = json_encode($user_json);
@@ -190,11 +188,14 @@ class BookRequestController extends Controller
     public function destroy($id)
     {
         $req = BookRequest::find($id);
-        $bookname = $req->book->name;
-        $req->delete();
-
-        Session::flash('success', '您已经取消了对'.$bookname.'的样书申请');
-        
+        $limit_result = $req->user->changeBookLimit(+1);
+        if($limit_result) {
+	        $req->delete();
+	        $bookname = $req->book->name;
+	        Session::flash('success', '您已经取消了对' . $bookname . '的样书申请');
+        }else{
+	        Session::flash('danger', '取消失败!');
+        }
         return redirect()->route('bookreq.record');
     }
 }
